debug css toolsSlider en responsive +amélioration ouverture de menu
