+++ src/datas/aboutMe.json
@@ -22,7 +22,7 @@
                 "twitter": "https://x.com/mariaintweb",
                 "linkedIn": "https://www.linkedin.com/in/maria-madalena-semedo-gon%C3%A7alves-033334336/",
                 "gitHub": "https://github.com/MMSGIWEB",
-                "name": "Semedo Gonçalves Maria Madalena",
+                "name": "Maria Semedo",
                 "email": "mariaintweb@gmail.com",
                 "location": "93160, Noisy-le-Grand, France"
             }
