+++ src/styles/sass/layouts/_Navbar.scss
@@ -10,8 +10,6 @@ header {
 
     .logo {
         width: 15%;
-        position: relative;
-        top: 1em;
     }
 
     nav {
@@ -22,7 +20,7 @@ header {
 
         .link {
             @include flexcc;
-            // @include glassEffect;
+
 
             &:active {
                 @include blackUnderline;
@@ -30,11 +28,12 @@ header {
 
             &:hover {
                 @include blackUnderline;
+                color: #000;
             }
 
             div {
-                color: #fff;
-                font-family: "DM Serif Display", serif !important;
+                color: #000;
+                font-family: "Roboto", sans-serif !important;
                 font-weight: 500;
                 text-wrap: nowrap;
                 font-size: 2em;
