+++ src/styles/sass/layouts/_cards.scss
@@ -9,6 +9,7 @@
     @include flexCCC;
 
     h1 {
+        margin-top: 0;
         font-size: 2em;
     }
 
@@ -40,7 +41,7 @@
                     @include flexcc;
 
                     .cardImg {
-                        height: 72px;
+                        height: 62px;
                         object-fit: fill;
                         border-radius: 10px;
                         width: 180px;
