+++ src/styles/sass/layouts/_slider.scss
@@ -49,6 +49,6 @@
         font-size: 1.5em;
         position: absolute;
         bottom: -40px;
-        font-family: 'Montserrat';
+        font-family: 'Roboto', sans-serif;
     }
 }
